requirements.txt
streamlit
openai
pandas
sqlalchemy
chromadb
sentence-transformers
optimum[openvino,nncf]
transformers
scikit-learn
python-dotenv
openpyxl
xlrd
plotly
psutil
httpx
reportlab
Pillow
kaleido
matplotlib
python-docx
pymupdf
mkl
mkl-service